feat: implementa a tela de denúncia com UI e navegação
